extract into packages and create handler to populate ES
