+++ .gitignore
@@ -1,0 +1,2 @@
+canonical-model
+data/
